requirements.txt
browser-use
fastapi
uvicorn[standard]
selenium
selenium-wire
python-dotenv
requests
playwright

